+++ docs/guide/usage.md
@@ -110,7 +110,7 @@ export default {
           presets: ['EXTERNAL_GEOMETRY', 'CONTEXT_DEPENDENT'], // [!code ++]
         }], // [!code ++]
         '@morev/bem/no-chained-entities': true, // [!code ++]
-        '@morev/bem/no-misplaced-side-effects': true, // [!code ++]
+        '@morev/bem/no-misplaced-relational-styles': true, // [!code ++]
         '@morev/bem/no-side-effects': true, // [!code ++]
         '@morev/bem/selector-pattern': true, // [!code ++]
       }, // [!code ++]
@@ -203,7 +203,7 @@ export default {
           ignoreBlocks: ['*swiper*'],
         }],
         '@morev/bem/no-chained-entities': [true, {}],
-        '@morev/bem/no-misplaced-side-effects': [true, {}],
+        '@morev/bem/no-misplaced-relational-styles': [true, {}],
         '@morev/bem/no-side-effects': [true, {
           ignore: [
             '*swiper*',
