+++ src/rules/bem/no-misplaced-relational-styles/__tests__/no-misplaced-relational-styles.configuration.test.ts
@@ -1,4 +1,4 @@
-import rule from '../no-misplaced-side-effects';
+import rule from '../no-misplaced-relational-styles';
 
 const { ruleName } = rule;
 const testRuleConfig = createTestRuleConfig({ ruleName });
