+++ src/rules/index.ts
@@ -4,7 +4,7 @@ export { default as blockVariable } from './bem/block-variable/block-variable';
 export { default as matchFileName } from './bem/match-file-name/match-file-name';
 export { default as noBlockProperties } from './bem/no-block-properties/no-block-properties';
 export { default as noChainedEntities } from './bem/no-chained-entities/no-chained-entities';
-export { default as noMisplacedSideEffects } from './bem/no-misplaced-side-effects/no-misplaced-side-effects';
+export { default as noMisplacedRelationalStyles } from './bem/no-misplaced-relational-styles/no-misplaced-relational-styles';
 export { default as noSideEffects } from './bem/no-side-effects/no-side-effects';
 export { default as selectorPattern } from './bem/selector-pattern/selector-pattern';
